Completed exercises through ex42; ex45 is a work in progress.  ex45.py is currently a script to generate a class file for a variety of different dungeon room types.  Notes on what I think I'll do next are in the bottom of that file in the comments.

+++ ex35.py
@@ -5,7 +5,7 @@ def gold_room():
     
     next = (raw_input("> "))
     try:
-         how_much = int(next)
+         how_much = float(next)
     except:
          dead("Man, learn to type a number.")
